azure-function/requirements.txt
# Requirements for Azure Function - Fixed versions for consistent deployment
# Using pure Python packages where possible to avoid build issues
openai==1.3.5  # Using a specific version to avoid compatibility issues
httpx==0.27.0  # Pin to a compatible version before the proxies parameter was removed
azure-functions==1.17.0
python-dateutil==2.8.2
requests==2.31.0
certifi==2023.11.17
tqdm==4.66.1
typing-extensions==4.8.0
charset-normalizer==3.3.2
idna==3.6
urllib3==2.0.7  # Added for OpenAI compatibility
# Removed aiohttp to avoid build issues